requirements.txt
beautifulsoup4==4.10.0
pandas==2.2.3
scikit-learn
ai21==2.15.1
uvicorn==0.31.0
fastapi==0.115.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
